ai/requirements.txt
fastapi
uvicorn[standard]
rapidfuzz
requests
google-generativeai==0.7.2
